Add Gemfile hack
Unfortunately bundler (1.x and 2.x) seems to be horrible at resolving
the dependencies here.

I chatted with someone who knows a fair bit more than me about how
bundler's algorithm works, and he suggested that the issue was that it
was having trouble locking down to a specific Rails version.

This allows us to specify the Rails version we want ("~> 5.1.0" does the
trick) which causes bundler to not struggle for 15 minutes to resolve
dependencies anymore.