Use spree address2 if it is present
The taxjar validation API has the option to take a freeform
street input consisting of the full address. This means that if both
address1 and address2 exist, we should concatenate them and pass them
to the street parameter.

This can help with situations that involve a suite number that has a
unique ZIP+4 code that can be determined from the suite. The suite
number is usually kept in address2, so without it, there's no way to
distinguish the full address if only the 5 digit zip code is passed.

Co-authored-by: Nick Van Doorn <nick@super.gd>