Add install generator
Prior to this commit, users were required to manually configure this
extension after installing the gem. We add an install generator so users
can run a single command to handle this.

We include this command in the sandbox script so the sandbox is
ready to use this extension by default.

Co-authored-by: Noah Silvera <noah@super.gd>