Pass line items to taxjar refund API
When a order transaction is created, the line items, their discount, and
sales tax is passed to the taxjar API.

For consistency and better reporting, we should pass the line items
relevant to a refund when calling the taxjar refund endpoint.
After investigation, we learned that a refund transaction is essentially
a delta of changes to the order, so the line items passed should reflect
only the line items the reimbursement was issued for, and the amounts
on those line items should be negative to reflect the refund delta.

One consequence of this is that the "discount" value does not need to be
passed, as the amount of discount on the original order has not changed.