Add transaction ID generator class
This will be used to generate transaction IDs for orders which we send
to the TaxJar reporting API. Orders can change after they have been sent
to TaxJar and the recommended way to update them is to cancel the
original order transaction and re-create it with the changes, however
the only way to associate the two transactions is through the
`transaction_id` field. This utility class implements the algorithm for
us to be able to do that by just appending a suffix and incrementing it
on each update.

Co-authored-by: Benjamin Willems <benjamin@super.gd>
Co-authored-by: Senem Soy <senem@super.gd>
Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Sam Lee <sam@super.gd>
Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>
Co-authored-by: Nick Van Doorn <nick@super.gd>