Only require state for Canada and US addresses
According to the TaxJar API, the `to_state` field is only required for
CA or US addresses. Right now, we consider an address incomplete if it
does not have a state, which breaks calculation of EU taxes as it is
commonplace to not have a "state"-equivalent.

We should change the incomplete check to only require the state if
TaxJar requires it for that country. This will allow calculation of tax
for EU addresses.

Co-authored-by: Madeline Collier <madeline@super.gd>