Inject `transaction_id` when building API params
Currently we always take the `order.number` as the `transaction_id`.
This is not going to be the case when we are creating an "updated"
transaction after changes have been made to an order and we will want to
use this field to associate the new transaction with the original. In a
subsequent change we will use this ability to specify a custom
`transaction_id` based on the context.

Co-authored-by: Benjamin Willems <benjamin@super.gd>