Support creating refund transactions from order
As we continue to add reporting to `solidus_taxjar` we need a method to
create refund transactions. Refund transactions look like the original
transaction but have all the payment amounts negated.

We use the TaxJar order instead of the Spree order because the Spree
order may have updated since the order was submitted to TaxJar.

Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Senem Soy <senem@super.gd>
Co-authored-by: Kendra Chateau <kendra@super.gd>
Co-authored-by: Alex Blackie <alex@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>