Use OrderTransaction.latest_for in "show" API calls
We now persist an `OrderTransaction` record any time we create a transaction on
TaxJar. In this change we try and fetch that record and if it exists, we use the
`transaction_id` from it when trying to call the TaxJar show API.

If no transaction ID is currently associated with an order, we raise an
error. We will handle this edge case in a future commit.

Co-Authored-By: Chris Todorov <chris@super.gd>
Co-Authored-By: Mike Conlin <mike@super.gd>