Move the install generator to the correct folder
The install generator wasn't working for the dummy app because it wasn't
in the correctfolder. The change in the Rakefile fixes the namespace, and
these both changes together make the install generator actually run when
the dummy app is generated.

One side effect of this change is that now the dummy app uses the TaxJar
tax calculator, which instantiates the api client. That will fail if
there is no TAXJAR_API_KEY environment configured, to prevent this we
are switching from the` .fetch` syntax to indexing the `ENV` hash. The
result of this is we will move the failure to when we make an api call
instead of when we instantiate the client.

Co-authored-by: Chris Todorov <chris@super.gd>
Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>