Add basic extension acceptance test
The most basic test of this extension is "does it calculate tax for
orders"

Since the testing application now automatically configures the
extension, we can just test whether a basic checkout will calculate
tax with Taxjar.

Solidus will try and load the google apis font on a CI where no
default fonts are available, so we should tell VCR to ignore this
request.

Co-authored-by: Nick Van Doorn <nick@super.gd>