Create the configuration model
We're introducing our first configuration setting to our tax calculator,
so we needed a place to store that setting. Any additional settings
we'll create in the future can also be stored in this table. For now the
reporting feature is defaulted to false, but we might want to change
that in the future depending on what Taxjar prefers.

Co-authored-by: Nick Van Doorn <nick@super.gd>