Hide the reporting admin UI
We'll keep this hidden from the users till the reporting feature is
fully implemented. We've added reporting_ui_enabled setting to use for
feature flagging, and replaced the current reporting_enabled setting
with that. This is to avoid confusion with the actual setting from the
model, which we'll integrate in the next commit.

Co-authored-by: Nick Van Doorn <nick@super.gd>