Run CI build against current versions of Solidus
This removes the lock we have on the Solidus extensions Orb version
which previous to this commit was running test only on "older" versions
of Solidus which were 2.9 and 2.10. With this change to the Orb version
as well as the switch to the "run-tests" command we will be running tests
against master, 3.1, 3.0 and 2.11 which are the officially supported versions
of Solidus currently.

Co-authored-by: Adam Mueller <adam@super.gd>