Add `Reporting#refund_and_create_new_transaction`
If a TaxJar transaction already exists for a Solidus order and the order
totals have changed, we need to cancel (or refund) the existing TaxJar
transaction and create a new one with the latest Solidus order totals on
it.

If the refund fails for some reason, we *do not* want to create the new
transaction. So we're testing that a new transaction is not created in
that case.

We can use this service method to do that in a child commit.

Co-Authored-By: Alex Blackie <alex@super.gd>
Co-Authored-By: Mike Conlin <mike@super.gd>