Increment TaxJar order transaction ID on create
If a TaxJar order transaction already exists for a Solidus order, we
need to increment the transaction ID used. For example: If an order
transaction for Solidus order number R123 already exists in TaxJar with
the transaction ID "R123", the next transaction ID will be "R123-1".

As of this commit, we're checking for order transactions known to exist
and incrementing based on the existing transactions IDs.

Co-Authored-By: Mike Conlin <mike@super.gd>