Remove `RAILS_VERSION` from Gemfile
This seems to have been added to address an issue with Bundler which has
been fixed since version 2.0 and is no longer required. By removing this
we will let the bundler take the version of Rails defined in the
solidus gemspec. As a result of this change we lose the ability to
override the Rails version in development, however ensure Circle.ci will
use the correct version without having to explicitly specify that
through the ENV.

This change also unlocks the version of `rake` we specify as a dev
dependency as that needs to change based on the Rails version we're
running tests against.

Co-authored-by: Adam Mueller <adam@super.gd>