Remove explicit test for event action return value
In Solidus 3.2.0.alpha there is a new event bus which has a slightly different
API than the existing one in which event handlers return instances of
`Spree::Event::Firing` class. The test we are removing here relies on the
existing API, however it doesn't provide any value as event handlers should not
be called explicitly outside of the event bus, thus we can safely remove this
test in order to allow our test suite to be forward compatible with current
versions of Solidus.