Handle unimplemented reporting features
If Solidus TaxJar's reporting functionality is enabled, and an order is
shipped, the `Solidus::Event#shipment_shipped` event that fires causes
an error. This is because we were not rescuing from the `StandardError`
that notified end users that a reporting feature had not yet been
implemented.

This error was unexpected and not handled gracefully. Until the
reporting functionality has been completed, we want the
`#shipment_shipped` event to *do nothing* if it hits an unimplemented
feature's codepath.