Simplify `Reporting#show_or_create_transaction`
The first time that a TaxJar order transaction is created for an order,
we know that TaxJar won't already have a transaction on record. We still
want to defensively check for a foreign reference to the TaxJar order
transaction in the context of an ActiveJob, in case the job is run more
than once.

In a subsequent change, we'll be re-adding the `NotImplementedError` we
removed here in our refund flow.

Co-Authored-By: Adam Mueller <adam.mathys@gmail.com>
Co-Authored-By: Noah Silvera <noah@super.gd>
Co-Authored-By: Chris Todorov <chris@super.gd>