Define a configurable logger instance
We want to implement request logging in future commits. We need to offer
users a way to configure this logger, so we define it in the root
module.

We overwrite the `logger` getter method to defer retrieving the default
unti later, as the `Rails.logger` is not available at app boot time.

Co-authored-by: Chris Todorov <chris@super.gd>