Take adjustments into account in discount calculator
TaxJar requires the line item totals to match the order total when
posting transactions. If you had canceled an item, and submitted a
refund for the cancelled item, re-posting that order to TaxJar would
fail, as the line_item would have the promo_total deducted but not the
adjustment made to the line item when it was cancelled.

We need to ensure any adjustments that are not tax are included in
the line item discount.

Co-authored-by: Nick Van Doorn <nick@super.gd>
Co-authored-by: Andrew Stewart <andrew@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>