Add models to for transaction backfill event
We want to record the activity of a transaction backfill. A transaction
backfill could contain a large number of orders, and will be processed in
the background. Therefore, we want to persist the results to the
database for the user to view at their leisure.

It's possible some transaction sync's may fail, so we want to be able
to record information about these failures and display it to the user.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Nick Van Doorn <nick@super.gd>