Allow admins to use TaxJar tax codes
This is a part of the certification requirements:

"Preferred: Users should be able to select a category using a dropdown,
“type-ahead”, or saved-search functionality synced with v2/categories
either manually or on a recurring basis.

Minimum: Provide a field for users to enter a product_tax_code and link
to our v2/categories datatable reference."

We have implemented the preferred method. We use a type-ahead dropdown
to sync TaxJar codes to existing Solidus tax codes. We provided a new
update path for the settings controller so that we only update what we
need to.

Co-authored-by: Benjamin Willems <benjamin@super.gd>