Remove order transaction creation from API class
Prior to this change, our API class was responsible for making API calls
to TaxJar and recording a local record of the transaction. We move
creating the transaction into the reporting class which is responsible
for integrating the TaxJar API with our data model.

This change prepares us for batch transaction syncing.