Move backfill button to backfill index page
Prior to this commit, the backfill button was on the already crowded
TaxJar settings page. Now that the backfill page is ready, we can move
it.

We also delete the old feature spec for starting a backfill and
implement a new one as we have much more information to inspect with
the new backfill UI.

Co-authored-by: Noah Silvera <noah@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>