Don't create new transactions for full refunds
If the entire amount of an order is refunded, and the new order total
is zero, we should not create a subsequent order transaction for zero
dollars after refunding.

We still want to allow some intentionally discounted zero dollar orders
be reported, as the item total affects the nexus and tax
calculations for certain states.

Co-authored-by: Chris Todorov <chris@super.gd>