Introduce a caching layer over the TaxJar API
This will be used for caching nexus regions initially but could be
expanded to other calls in the future. The reason why we want to cache
nexus regions is because now we are performing the nexus region check before
each tax calculation and we don't want perform the call to fetch the
nexus regions every time.

Co-authored-by: Noah Silvera <noah@super.gd>