Resolve test failures
In this commit we:

- Re-record cassettes to account for new requests for nexus regions.
- Adjust tests that never cared about nexus region checks before now.

Now that we've introduced an additional API call for each tax
calculation, we need to re-record many cassettes to include the request
and response checking for active nexus regions.

In the `taxjar_settings_spec`, we can no longer assert that "British
Columbia" is not on the page before we sync nexus regions because of our
new nexus regions cache layer.

Co-authored-by: Chris Todorov <chris@super.gd>
Co-authored-by: Benjamin Wil <benjamin@super.gd>