Update spec setup to reset state in `before` block
In a recent change to core, an `after(:build)` hook was added to set the
state on the address record for countries which require states. This
interferes with the spec setup by ignoring the state attribute we were
setting to `nil`. Instead we have moved that to a `before` block so that
we can correctly setup the test address to not have a state.

Relevant change upstream - https://github.com/solidusio/solidus/pull/4272

Co-authored-by: Andrew Stewart <andrew@super.gd>