Fix broken negative tests in reporting subscriber
Previously, these tests were asserting that a specific method within
a job did not get called.

This a.) broke encapsulation of the test by relying on details of how
the job works and b.) wasn't a good negative test, as the job runs
asynchronously, and we were ensuring the job was run before the test
completed.

Instead of testing the internals of the job to check if it ran, we should
just check if it got queued.

Co-authored-by: Nick Van Doorn <nick@super.gd>