Remove `order_recalculated` event backport
We no longer officially support Solidus 2.9 and 2.10 on the `master`
branch of this extension and in the upcoming 1.0.0 release, since we
stopped running tests on CI against those versions here -
https://github.com/SuperGoodSoft/solidus_taxjar/pull/135. This is also
inline with the official Solidus policy of supporting only 3 releases
back.

With that in mind we can stop maintaining this backported event, which
is needed for transaction reporting to work on Solidus 2.9 and 2.10.