Report `payment_total` to TaxJar
Prior to this commit, we reported `order.total` to TaxJar which in some
cases does not reflect the total liability transferred by the customer.
For example, issuing reimbursements does not change the total. To fix
this, we report `order.payment_total` to TaxJar as it accurately
reflects the liability transfer.