Move order pre-conditions into `with_reportable`
These conditions were already implicitly true for the
`report_transaction` event action, so we can safely move them into the
shared module which now provides all the checks that need to be
satisfied by an order.

The tests needed a small change to better reflect the state of the
shipment and the order, when the `shipment_shipped` event is fired (i.e.
after the transition of a shipment to the `shipped` state).

Co-authored-by: Benjamin Willems <benjamin@super.gd>