Move `transaction_replaceable?` to module
This is the last step of the refactoring work to extract the logic
around `Reportable` orders. This change combines the conditions for
whether a transaction on TaxJar is replaceable into a single public API
we can expose on the `Reportable` module. This will allow us to
condition logic on this check by simply including this module wherever
we need to re-use this logic.