Fire `shipment_shipped` after running `after_ship`
For some reason our feature spec will receive a shipment record which
isn't yet shipped, and that is likely because in our feature specs we
run the jobs for the reporting subscriber async, unlike in our unit
tests where we just assert that a job has been queued.

Firing the event is done synchronously, so this change will ensure that
the service layer that handles generating Carton objects when inventory units
# are actually shipped and also updates the order and shipment states
has been performed (in `Spree::OrderShipping`), before the event is
consumed by our subscriber.

We tried updating the unit specs to expose this, but we were not successful.

Co-authored-by: benjamin wil <benjamin@super.gd>