Load factories in line with new Solidus conventions
This change updates our extension to use the new method provided by
`solidus_dev_support` instead of manually requiring the files. This is
in line with how `FactoryBot` recommends factories are registered and
resolves an issue with our custom override for the `Spree::Address`
factory.

Currently the upstream version of `solidus_dev_support` does not handle
our use case of having a two folder deep namespace, so this change needs
to wait for that to be fixed. We have opened a PR upstream for that to
be changed - https://github.com/solidusio/solidus_dev_support/pull/207

Co-authored-by: Nick Van Doorn <nick@super.gd>