Split job for running tests against current/older Solidus
We now need to bump the version of Ruby used to run specs against
"current" versions of Solidus because Solidus 4.0 is out and it requires
Ruby 3.0. Previously in the same job we also ran the specs against older
versions of Solidus.

This change splits current and older into separate jobs and gives them
different versions of Ruby. This also updates the version of Ruby used
to run specs against the main branch to be '3.1'.