Rename current subscriber to Legacy and add Omnes version
In order to support the new Omnes event system on Solidus 4.x, we are
duplicating our existing subscriber class and renaming it to
`LegacyReportingSubscriber` and updating the implementation of what will
be the new `ReportingSubscriber` to use the new Omnes API.

One change we realized is necessary is to move both subscriber files
to the `lib` folder so they are no longer automatically loaded by the host app.
This means we need to manually require and "activate" the correct
subscriber based on the Solidus version. For this we are introducing a
`solidus_taxjar_legacy_events.rb` initializer for versions of Solidus
where Omnes has not been introduced yet - prior to 3.2.

Co-authored-by: Cameron Day <cameron@super.gd>
Co-authored-by: benjamin wil <benjamin@super.gd>