Prevent order from being reported if it is not paid
In manual testing, we ran into some scenarios where we want an order to
be reported but it can end up in a shipped and unpaid state.

e.g. If you have an address that caused a failed sync to taxjar due to
validation errors, and then
correct the address, the tax on the order may change, resulting
 in an order that has never been reported to taxjar, but is unpaid and
 shipped.

 Co-authored-by: Chris Todorov <chris@super.gd>