Combine the replace and report event handlers
Because we now need a way to recover from failures to report
transactions after updates to them happen, we have decided to combine
the event handlers into one in order to avoid having to manage multiple
flows for the same action. This will also allow us to remove the need
for the custom `shipment_shipped` event this extension needs to
maintain.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>