Run `bundle exec solidus extension .` to update binstubs
This fixes the `bin/sandbox` when running with SOLIDUS_BRANCH=`main`.
Previous to these changes, the building of the sandbox app would fail
with an error finding `/config/routes.rb` during the running of the
`solidus:install` task as part of the script. It's not immediately clear
which of these changes fixed the issue, but this is what
`solidus_dev_support` generates now for new extensions, so let's run
with it.

Co-authored-by: Alistair Norman <alistair@super.gd>